pyproject.toml
[project]
name = "ComfyUI-Flux-TryOff"
description = "Try Off for ComfyUI using Flux and CatVTON. Now with Quantized (8-bit) Support."
version = "1.2.0"
license = ""
dependencies = ["numpy", "pillow", "diffusers", "accelerate", "transformers"]

[project.urls]
Repository = "https://github.com/asutermo/ComfyUI-Flux-TryOff"
#  Used by Comfy Registry https://comfyregistry.org

[tool.comfy]
PublisherId = "asutermo"
DisplayName = "ComfyUI-Flux-TryOff"
Icon = ""